packages/detektor/detektor-1.1.0-beta.012.tar.gz/detektor-1.1.0-beta.012/detektor/languageparser/python.py
from detektor.codefilter.normalizeoperatorwhitespace import NormalizeOperatorWhitespace
from detektor.languageparser.shlexparser import ShlexLanguageParserBase
import keyword


class PythonLanguageParser(ShlexLanguageParserBase):
    operators = {
        '!', '+', '-', '*', '**', '/', '//', '%', '<<', '>>', '&', '|',
        '=', '%=', '/=', '//=', '**=', '*=', '-=',
        '^', '~', '<', '>', '<=', '>=', '==', '!=', '<>',
    }
    keywords = set(keyword.kwlist)
    sourcecode_preprocessor_classes = [
        NormalizeOperatorWhitespace
    ]
